% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Title}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{data frame with two variables, w and v.}

\item{W}{integer containg the maximum weight of our knapsack}
}
\value{
list of two lists, value showing the max value of our optimal knapsack and the elements of that knapsack
}
\description{
Title
}
